requirements.txt
requests==2.25.1
numpy==1.20.2
matplotlib==3.4.1
pillow>=8.3.2
PyYAML==5.4.1
#rpi.gpio==0.7.0
nh-currency==1.0.1
